refactor: moved some message component to the correct module directory
