+++ src/components/Messages/components.ts
@@ -62,8 +62,8 @@ import MsgMultiSendComponents from './bank/MsgMultiSend';
 import MsgVoteComponents from './gov/MsgVote';
 import MsgDelegateComponents from './staking/MsgDelegate';
 import MsgWithdrawDelegatorRewardsComponents from './distribution/MsgWithdrawDelegatorRewards';
-import MsgBeginRedelegateComponents from './distribution/MsgBeginRedelegate';
-import MsgUndelegateComponents from './distribution/MsgUndelegate';
+import MsgBeginRedelegateComponents from './staking/MsgBeginRedelegate';
+import MsgUndelegateComponents from './staking/MsgUndelegate';
 import MsgSaveProfileComponents from './profiles/MsgSaveProfile';
 import MsgLinkChainAccountComponents from './profiles/MsgLinkChainAccount';
 import MsgUnlinkChainAccountComponents from './profiles/MsgUnlinkChainAccount';
