+++ src/components/Messages/staking/MsgBeginRedelegate/MsgBeginRedelegateDetails/index.tsx
